+++ src/getpodia/course.nim
@@ -14,6 +14,9 @@ from std/xmltree import findAll, attrs, innerText, XmlNode, items, attr
 export tables.`[]`
 
 import pkg/findxml/findAll
+import pkg/scraper
+
+from pkg/util/forStr import between
 
 type
   Course* = object
@@ -23,7 +26,7 @@ type
     name*: string
     videos*: seq[CourseVideo]
   CourseVideo* = object
-    name*, pageUrl*, code*: string
+    name*, pageUrl*, code*, fileUrl*: string
     comments*: seq[VideoComment]
     meta*: VideoMeta
   VideoComment* = ref VideoCommentObj
@@ -64,7 +67,7 @@ proc extractCourse*(client; url: string): Course =
     html = parseHtml client.getContent url
   for lect in html.findAll("li", {"class": "list-group list-group-menu list-group-xs mb-6"}):
     var lecture: CourseLecture
-    lecture.name = lect.findAll("h2", {"class": "h4 text-md mb-2"})[0].innerText.strip
+    lecture.name = lect.findAll("h2", {"class": "h4 text-md mb-2"}).text.strip
     for vidEl in lect.findAll("a", {"class": "list-group-item list-group-item-action"}):
       var video: CourseVideo
       # video.pageUrl = $(urlBase / vidEl.attrs["href"])
@@ -72,12 +75,12 @@ proc extractCourse*(client; url: string): Course =
       video.name = vidEl.innerText.strip
       lecture.videos.add video
     result.lectures.add lecture
-  result.name = html.findAll("h1", {"class": "h4"})[0].innerText.strip
+  result.name = html.findAll("h1", {"class": "h4"}).text.strip
   block description:
-    let descEl = html.findAll("div", {"class": "mt5 text-xs"})
+    let descEl = html.findAll("div", {"class": "text-muted text-lg mt-n4 spacer-section"})
     if descEl.len == 0:
       break
-    result.description = descEl[0].innerText.strip
+    result.description = descEl.text.strip
   result.image = html.findAll("a", {"class": "thumb thumb-lg thumb-photo thumb-link"})[0].attrs["style"][22..^3]
   result.code = url.getCode
 
@@ -92,42 +95,42 @@ proc update*(client; self: var CourseVideo) =
   let
     body = client.getContent self.pageUrl
     html = parseHtml body
-  self.name = html.findAll("h1", {"class": "pr-md-6 d-md-block d-none"})[0].innerText.strip
+  self.name = html.findAll("h1", {"class": "pr-md-6 d-md-block d-none"}).text.strip
   block videoCode:
-    var startI = body.findText("_wq.push({") - 1
-    var json = body[startI..^1]
-    var endI = json.find("});") + 1
-    json = json[0..<endI]
-
-    startI = json.findText "id: \""
-    self.code = json[startI..^1]
-    endI = self.code.find("\"") - 1
-    self.code = self.code[0..endI]
-  # block comments:
-  #   proc getCommentData(node: XmlNode): VideoComment =
-  #     new result
-  #     echo node
-  #     result.avatar = node.findAll("div", {"class": "avatar avatar-xs"})[0].attrs["style"][22..^3]
-  #     result.author = node.findAll("a", {"class": "font-weight-bold text-sm text-grey-dark my-1"})[0].innerText.strip
-  #     # result.body = node.findAll([
-  #     #   ("div", @{"class": "col-sm-12"}),
-  #     #   ("div", @{"class": "ml2 text-xs"})
-  #     # ])[0].innerText.strip
-  #     result.body = node.findAll("div", {"class": "text-longform mb-3"})[0].innerText.strip
-  #     result.creation = node.findAll("span", {"class": "text-muted text-sm my-1"})[0].attrs["data-title"]
-  #     let likes = node.findAll("a", {"class": "btn btn-icon btn-transparent-soft btn-xs btn-rounded btn-tight"})[0].innerText.strip
-  #     if likes.len > 7:
-  #       result.likes = likes[7..^1].parseInt
-  #   proc getNested(comment: VideoComment; commentNode: XmlNode) =
-  #     for tdiv in commentNode.findAll("div"):
-  #       if "replies" in tdiv.attr "id":
-  #         for nest in tdiv:
-  #           comment.nested.add nest.getCommentData
-
-  #   for com in html.findAll("div", {"data-controller": "podia--comments--comment--form-component"}):
-  #     let comment = com.getCommentData
-  #     comment.getNested com
-  #     self.comments.add comment
+    self.code = body.between("_wq.push({", "});").between("id: \"", "\"")
+    if self.code.len == 0:
+      self.fileUrl = html.findAll([
+        ("div", @{"id": "lesson-content"}),
+        ("a", @{"class": "list-group-item list-group-item-action"})
+      ]).attr "href"
+      if "podia.com" notin self.fileUrl:
+        var link = parseUri self.pageUrl
+        link.path = self.fileUrl
+        self.fileUrl = $link
+  block comments:
+    proc getCommentData(node: XmlNode): VideoComment =
+      new result
+      result.avatar = node.findAll("div", {"class": "avatar avatar-xs"})[0].attrs["style"][22..^3]
+      result.author = node.findAll("a", {"class": "font-weight-bold text-sm text-grey-dark my-1"}).text.strip
+      result.body = node.findAll("div", {"class": "text-longform mb-3"}).text.strip
+      result.creation = node.findAll("span", {"class": "text-muted text-sm my-1"})[0].attrs["data-title"]
+      let likes = node.findAll("a", {"class": "btn btn-xs btn-transparent-soft btn-rounded btn-tight btn-icon"}).text.strip
+      if likes.len > 7:
+        result.likes = likes[7..^1].parseInt
+    proc getNested(comment: VideoComment; commentNode: XmlNode) =
+      for tdiv in commentNode.findAll("div"):
+        if "replies" in tdiv.attr "id":
+          for nest in tdiv:
+            try:
+              comment.nested.add nest.getCommentData
+            except: discard
+              
+    for com in html.findAll("div", {"data-controller": "podia--comments--comment--form-component"}):
+      if com.attr("id") == "comments":
+        continue
+      let comment = com.getCommentData
+      comment.getNested com
+      self.comments.add comment
 
 proc url*(self: CourseVideo): VideoSource =
   VideoSource("https://fast.wistia.net/embed/iframe/" & self.code)
@@ -140,10 +143,7 @@ proc getMeta*(self: var CourseVideo) =
   })
   let body = client.getContent self.url.string
 
-  let startI = body.findText "W.iframeInit("
-  var jsonData = body[startI..^1]
-  let endI = jsonData.find ", {});"
-  jsonData = jsonData[0..<endI]
+  var jsonData = body.between("W.iframeInit(", ", {});")
   let json = parseJson jsonData
   self.meta.filename = json{"name"}.getStr
   for asset in json{"assets"}:
@@ -172,20 +172,20 @@ when isMainModule:
   from std/tables import pairs
   from std/json import `%*`, pretty
   import std/jsonutils
-  when false:
+  when true:
     var video = CourseVideo(
-      pageUrl: "http://localhost:5555/.test/videos/A%20Doa%C3%A7%C3%A3o%20de%20sua%20Hist%C3%B3ria.html" # Some downloaded Podia video page
+      pageUrl: "http://127.0.0.1:5555/.test/videos/PDF%20-%20Prece%20contra%20Esp%C3%ADrito%20do%20Ci%C3%BAmes.html" # Some downloaded Podia video page
     )
     newHttpClient().update video
     # video.getMeta
-    echo video
-    for name, m in video.meta.medias:
-      echo name, %*m
-    echo "---"
-    echo video.meta.thumbnail
-    echo video.meta.hdVideo
-    echo "---"
-    echo pretty video.comments.toJson
+    # echo video
+    # for name, m in video.meta.medias:
+    #   echo name, %*m
+    # # echo "---"
+    # echo video.meta.thumbnail
+    # echo video.meta.hdVideo
+    # echo "---"
+    echo pretty video.toJson
     # for comment in video.comments:
     #   echo comment[]
     #   for comment in comment.nested:
@@ -193,5 +193,5 @@ when isMainModule:
     #     for comment in comment.nested:
     #       echo "    ", comment[]
   else:
-    let course = newHttpClient().extractCourse "http://localhost:5555/.test/courses/A%20Arte%20da%20Comunica%C3%A7%C3%A3o_%20Storytelling%20para%20Amizades%20e%20Neg%C3%B3cios.html"
+    let course = newHttpClient().extractCourse "http://127.0.0.1:5555/.test/courses/BATALHA%20ESPIRITUAL%20I%20Entendendo%20o%20Advers%C3%A1rio.html"
     echo pretty course.toJson